# server_python_e_client_java/soseng-project-airline/python_server/test_esempi/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: compagniaAereaAPI
  description: È l'API Restful offerta dalla compagnia aerea che permette di fruire
    alla capability di restituire le offerte attive e di riceve la quota del pagamento
    del cliente.
  version: "1.0"
servers:
- url: https://stoplight.io/mocks/soseng-unibo/soseng-project-documentation/7424067
  description: testing
paths:
  /flights:
    get:
      tags:
      - Risorse
      summary: Restituisci le offerte attive
      description: È la risorsa che restituisce tutte le offerte di voli attive.
      operationId: get_flights
      parameters: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
              examples:
                Esempio:
                  value:
                    flights:
                    - departure-from: BLQ
                      takeoff: "2020-04-27, 10.00AM, CET"
                      destination: BCN
                      price:
                        value: 100
                        currency: €
                      offer_code: ryan253
                    companyname: ryanflights
        "500":
          description: Internal Server Error
      x-openapi-router-controller: swagger_server.controllers.risorse_controller
  /LMflight:
    summary: Crea un volo last minute
    description: "Endpoint che permette di creare un volo lastminute, specificando\
      \ tutti i parametri del volo; il server lo inoltrerà poi ad acmesky"
    post:
      tags:
      - Risorse
      summary: crea un volo last minute
      operationId: post_lmflight
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/lmflight'
      responses:
        default:
          description: Default error sample response
        "200":
          description: OK
        "400":
          description: CLIENT ERROR
        "500":
          description: SERVER ERROR
      security:
      - companytoken: []
      x-openapi-router-controller: swagger_server.controllers.risorse_controller
  /notifypayment:
    post:
      tags:
      - Risorse
      summary: Ricevi pagamento
      description: È la risorsa che permette al fornitore dei servizi bancari di inviare
        alla compagnia aerea la quota del pagamento ricevuto dal cliente per l'acquisto
        dell'offerta.
      operationId: post_notifypayment
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/notifypayment_body'
      responses:
        "200":
          description: OK
      security:
      - companytoken: []
      x-openapi-router-controller: swagger_server.controllers.risorse_controller
  /registration:
    post:
      tags:
      - Risorse
      summary: Registra un nuovo utente
      description: È la risorsa che permette a chi intende interagire con la compagnia
        aerea di ottenere l'APIKey per poter essere identificato e autorizzato.
      operationId: post_registration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Maps.v1_Credentials'
            examples:
              Esempio:
                value:
                  username: Mario
                  password: 12a4a54s
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
              examples:
                Esempio:
                  value:
                    token: bWFnZ2llOnN1bW1lcnM=
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      x-openapi-router-controller: swagger_server.controllers.risorse_controller
components:
  schemas:
    Maps.v1_Credentials:
      title: credentials
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    inline_response_200:
      type: object
      properties:
        flights:
          uniqueItems: false
          type: array
          items:
            $ref: '#/components/schemas/LMflight_flight'
        companyname:
          type: string
      example:
        companyname: companyname
        flights:
        - takeoff: takeoff
          price:
            amount: 0.8008281904610115
            currency: currency
          destination: destination
          offer_code: offer_code
          departure-from: departure-from
        - takeoff: takeoff
          price:
            amount: 0.8008281904610115
            currency: currency
          destination: destination
          offer_code: offer_code
          departure-from: departure-from
    LMflight_flight_price:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
      example:
        amount: 0.8008281904610115
        currency: currency
    LMflight_flight:
      type: object
      properties:
        departure-from:
          type: string
        takeoff:
          type: string
        destination:
          type: string
        price:
          $ref: '#/components/schemas/LMflight_flight_price'
        offer_code:
          type: string
          description: Codice identificativo dell'offerta
      example:
        takeoff: takeoff
        price:
          amount: 0.8008281904610115
          currency: currency
        destination: destination
        offer_code: offer_code
        departure-from: departure-from
    lmflight:
      title: lmflight
      type: object
      properties:
        flight:
          $ref: '#/components/schemas/LMflight_flight'
        companyname:
          type: string
    notifypayment_customer:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
    notifypayment_amount_payed:
      type: object
      properties:
        value:
          type: number
        currency:
          type: string
    notifypayment_transaction:
      type: object
      properties:
        date:
          type: string
        id:
          type: number
    notifypayment_body:
      type: object
      properties:
        offer_code:
          type: string
          description: Codice identificativo dell'offerta
        customer:
          $ref: '#/components/schemas/notifypayment_customer'
        amount_payed:
          $ref: '#/components/schemas/notifypayment_amount_payed'
        transaction:
          $ref: '#/components/schemas/notifypayment_transaction'
    inline_response_200_1:
      type: object
      properties:
        token:
          type: string
      example:
        token: token
  securitySchemes:
    companytoken:
      type: apiKey
      description: Schema di sicurezza per permettere solo ai clienti opportunatamente
        autenticati di inviare il pagamento.
      name: authtoken
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_companytoken